chore: split Eleventy configuration by responsibility
Move concern-specific filters, shortcodes, collections, image handling, assets, and search registration into focused modules while keeping the root configuration as the assembly point. Document module ownership and include the new configuration path in CI triggers.
