+++ docs/README.md
@@ -22,7 +22,7 @@ Use this page as the entry point for repository documentation. It maps each doc 
 | [`../worker/README.md`](../worker/README.md) | Feedback worker routes, local dev, and deploy commands | Maintainer/operators | You are operating or debugging `feedback.allaboutken.com` |
 | [`../src/site/style-guide/editorial.njk`](../src/site/style-guide/editorial.njk) | Canonical voice/tone/style guidance | Authors and editors | You need writing standards while drafting or editing |
 | [`FRONTMATTER.md`](FRONTMATTER.md) | Reference for every YAML frontmatter field used across post, page, and digest templates | Authors, editors, and contributors adding posts | You need to know what a field does, which templates use it, and whether it is required |
-| [`ELEVENTY_CONFIG.md`](ELEVENTY_CONFIG.md) | Reference for all plugins, filters, shortcodes, collections, and config options in `eleventy.js` | Contributors modifying build config or templates | You are adding a filter, shortcode, or collection, or need to understand how the build pipeline works |
+| [`ELEVENTY_CONFIG.md`](ELEVENTY_CONFIG.md) | Reference for all plugins, filters, shortcodes, collections, and config options assembled by `eleventy.js` and `config/eleventy/` | Contributors modifying build config or templates | You are adding a filter, shortcode, or collection, or need to understand how the build pipeline works |
 | [`DATA_FILES.md`](DATA_FILES.md) | Schema and template usage for `siteConfig.json` and `garage.json` | Contributors editing site-wide config or the garage project listing | You are adding a garage project, updating site metadata, or tracing where a global value comes from |
 | [`TEMPLATES.md`](TEMPLATES.md) | Layout hierarchy, partials, macros, and Nunjucks patterns used in site templates | Contributors changing page/layout templates | You are adding a page template, editing layouts, or tracing where markup is shared |
 | [`COMPONENTS.md`](COMPONENTS.md) | Component directory structure and Sass/asset build wiring | Contributors editing UI components and styles | You are adding/changing SCSS modules, assets, or component markup |
